# rust-services/openapi.yaml
openapi: 3.0.3
info:
  title: Penpot Rust Microservices API
  description: |
    High-performance microservices for Penpot design platform.
    
    ## Services
    - **API Gateway** (8080): Central routing and caching
    - **Shape Validator** (8081): Shape validation
    - **Realtime Sync** (8082): WebSocket collaboration
    - **Render Service** (8083): SVG rendering
  version: 0.1.0
  license:
    name: MPL-2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/
servers:
  - url: http://localhost:8080
    description: API Gateway
  - url: http://localhost:8081
    description: Shape Validator
  - url: http://localhost:8082
    description: Realtime Sync
  - url: http://localhost:8083
    description: Render Service
...
paths:
...
  /api/v1/validate:
    post:
      summary: Validate shapes via gateway
      tags: [gateway]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateResponse'
        '503':
          description: Validator service unavailable
...
  /validate:
    post:
      summary: Validate shapes
      tags: [validation]
      servers:
        - url: http://localhost:8081
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateRequest'
            example:
              shapes:
                - id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "Rectangle"
                  type: "rect"
                  x: 0
                  y: 0
                  width: 100
                  height: 100
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateResponse'
              example:
                valid: true
                total_shapes: 1
                valid_shapes: 1
                invalid_shapes: 0
                total_errors: 0
                processing_time_us: 5

  # =========================================================================
  # RENDER SERVICE (8083)
  # =========================================================================
...
components:
  schemas:
    ValidateRequest:
      type: object
      required:
        - shapes
      properties:
        shapes:
          type: array
          items:
            $ref: '#/components/schemas/Shape'
    ValidateResponse:
      type: object
      properties:
        valid:
          type: boolean
        total_shapes:
          type: integer
        valid_shapes:
          type: integer
        invalid_shapes:
          type: integer
        total_errors:
          type: integer
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        processing_time_us:
          type: integer
    Shape:
      type: object
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [rect, circle, path, text, image, frame, group, bool, svg-raw]
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
    ValidationError:
      type: object
      properties:
        shape_id:
          type: string
          format: uuid
        message:
          type: string
...